requirements.txt
Pillow==7.2.0
pandas==1.2.0
session_info==1.0.0
streamlit==1.7.0
Keras-Preprocessing==1.1.2
numpy==1.19.5
tensorflow==2.4.1
google-cloud
google-cloud-bigquery
google-cloud-storage